requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dateutil
streamlit
pandas
